refactor(app): create separate folders and services for shipping api
