+++ back/src/shipping/dto/product-qty.dto.ts
@@ -1,7 +1,6 @@
 import { IsNotEmpty, IsNumber, Min} from "class-validator";
 
 export class ProductQtyDto {
-
     @IsNotEmpty()
     productId: number;
