+++ back/src/shipping/dto/shipping-list.response.ts
@@ -1,5 +1,5 @@
 import { ShipmentSummaryDto } from './shipment-summary.dto';
-import { PaginationDtoOut } from './pagination.dto';
+import { PaginationDtoOut } from '../../shared/dto/pagination.dto';
 
 export class ShippingListResponse {
   shipments!: ShipmentSummaryDto[];
