+++ back/src/shipping/dto/shipping-log.dto.ts
@@ -1,12 +1,7 @@
-import { ShippingStatus } from './shipping-status.enum';
+import { ShippingStatus } from '../../shared/enums/shipping-status.enum';
 
 export class ShippingLogDto {
-  /** Fecha y hora del cambio (ISO-8601, UTC) */
   timestamp!: string;
-
-  /** Nuevo estado del envío */
   status!: ShippingStatus;
-
-  /** Mensaje legible */
   message!: string;
 }
