+++ back/src/shipping/dto/transport.dto.ts
@@ -1,7 +1,6 @@
 import { IsNotEmpty, IsIn } from "class-validator"
 
 export class TransportDto {
-
     @IsNotEmpty()
     @IsIn(['air', 'sea', 'road', 'rail'])
     type: 'air' | 'sea' | 'road' | 'rail';
